@@ -2,10 +2,10 @@ import { Component, Input, OnInit } from '@angular/core';
 import { MimeViewerConfig } from '@nationallibraryofnorway/ngx-mime';
 
 @Component({
-    selector: 'nationallibraryofnorway-root',
-    templateUrl: './app.component.html',
-    styleUrls: ['./app.component.scss'],
-    standalone: false
+  selector: 'nationallibraryofnorway-root',
+  templateUrl: './app.component.html',
+  styleUrls: ['./app.component.scss'],
+  standalone: false,
 })
 export class AppComponent implements OnInit {
   @Input() manifestUri!: string;
