@@ -4,6 +4,7 @@ import {
   Component,
   OnDestroy,
   OnInit,
+  inject,
 } from '@angular/core';
 import { Subscription } from 'rxjs';
 import { IiifManifestService } from './../../core/iiif-manifest-service/iiif-manifest-service';
@@ -11,21 +12,18 @@ import { MimeViewerIntl } from './../../core/intl';
 import { Manifest } from './../../core/models/manifest';
 
 @Component({
-    selector: 'mime-metadata',
-    templateUrl: './metadata.component.html',
-    styleUrls: ['./metadata.component.scss'],
-    changeDetection: ChangeDetectionStrategy.OnPush,
-    standalone: false
+  selector: 'mime-metadata',
+  templateUrl: './metadata.component.html',
+  styleUrls: ['./metadata.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: false,
 })
 export class MetadataComponent implements OnInit, OnDestroy {
+  intl = inject(MimeViewerIntl);
+  private changeDetectorRef = inject(ChangeDetectorRef);
+  private iiifManifestService = inject(IiifManifestService);
   public manifest: Manifest | null = null;
   private subscriptions = new Subscription();
-
-  constructor(
-    public intl: MimeViewerIntl,
-    private changeDetectorRef: ChangeDetectorRef,
-    private iiifManifestService: IiifManifestService,
-  ) {}
 
   ngOnInit() {
     this.subscriptions.add(
