@@ -3,7 +3,13 @@ import {
   BreakpointState,
   Breakpoints,
 } from '@angular/cdk/layout';
-import { ChangeDetectorRef, Component, OnDestroy, OnInit } from '@angular/core';
+import {
+  ChangeDetectorRef,
+  Component,
+  OnDestroy,
+  OnInit,
+  inject,
+} from '@angular/core';
 import { Subscription } from 'rxjs';
 import { AltoService } from '../core/alto-service/alto.service';
 import { IiifManifestService } from '../core/iiif-manifest-service/iiif-manifest-service';
@@ -17,12 +23,20 @@ import { ViewerLayout } from '../core/models/viewer-layout';
 import { ViewerLayoutService } from '../core/viewer-layout-service/viewer-layout-service';
 
 @Component({
-    selector: 'mime-view-dialog',
-    templateUrl: './view-dialog.component.html',
-    styleUrls: ['./view-dialog.component.scss'],
-    standalone: false
+  selector: 'mime-view-dialog',
+  templateUrl: './view-dialog.component.html',
+  styleUrls: ['./view-dialog.component.scss'],
+  standalone: false,
 })
 export class ViewDialogComponent implements OnInit, OnDestroy {
+  intl = inject(MimeViewerIntl);
+  private breakpointObserver = inject(BreakpointObserver);
+  private cdr = inject(ChangeDetectorRef);
+  private viewerLayoutService = inject(ViewerLayoutService);
+  private iiifManifestService = inject(IiifManifestService);
+  private altoService = inject(AltoService);
+  private mimeResizeService = inject(MimeResizeService);
+
   tabHeight = {};
   isHandsetOrTabletInPortrait = false;
   viewerLayout: ViewerLayout = ViewerLayout.ONE_PAGE;
@@ -33,16 +47,6 @@ export class ViewDialogComponent implements OnInit, OnDestroy {
   RecognizedTextMode: typeof RecognizedTextMode = RecognizedTextMode;
   private mimeHeight = 0;
   private subscriptions = new Subscription();
-
-  constructor(
-    public intl: MimeViewerIntl,
-    private breakpointObserver: BreakpointObserver,
-    private cdr: ChangeDetectorRef,
-    private viewerLayoutService: ViewerLayoutService,
-    private iiifManifestService: IiifManifestService,
-    private altoService: AltoService,
-    private mimeResizeService: MimeResizeService,
-  ) {}
 
   ngOnInit(): void {
     this.subscriptions.add(
